Refactored date utils into date-utils.
