+++ apps/shoroukelkobrsi.com/src/components/film/film-showcase.tsx
@@ -6,6 +6,7 @@ import { unstable_cache as cache } from "next/cache";
 import configPromise from "@payload-config";
 import type { Film, Homepage } from "@/payload-types";
 import AspectRatio from "@/components/ui/aspect-ratio";
+import { formatSeasonYear } from "@/utilities/date-utils";
 import styles from "./film-showcase.module.css";
 
 const getCachedHomepage = cache(
@@ -25,33 +26,6 @@ const getCachedHomepage = cache(
     tags: ["homepage"],
   },
 );
-
-function getSeason(date: Date): string {
-  const month = date.getMonth();
-  switch (month) {
-    case 2:
-    case 3:
-    case 4:
-      return "Spring";
-    case 5:
-    case 6:
-    case 7:
-      return "Summer";
-    case 8:
-    case 9:
-    case 10:
-      return "Autumn";
-    default:
-      return "Winter";
-  }
-}
-
-function formatSeasonYear(dateString: string): string {
-  const date = new Date(dateString);
-  const season = getSeason(date);
-  const year = date.getFullYear().toString();
-  return `${season} ${year}`;
-}
 
 export default async function FilmShowcase(): Promise<React.ReactElement> {
   const homepage = await getCachedHomepage();
